refactor: resolve issues #32 points 2 & 3 - extract useScan and useOpenSubtitles hooks, create VideoCard component
